feat:delete initial page.
