+++ src/components/auth/login-form.tsx
@@ -1,0 +1,214 @@
+'use client';
+
+import { motion } from 'framer-motion';
+import { ArrowRight, Lock, Mail, Sparkles } from 'lucide-react';
+import Link from 'next/link';
+import { useRouter } from 'next/navigation';
+import { signIn, useSession } from 'next-auth/react';
+import { useId, useState } from 'react';
+import Memo from '~/app/_components/memo';
+import { Button } from '~/components/ui/button';
+import { Card, CardContent, CardHeader, CardTitle } from '~/components/ui/card';
+import { Input } from '~/components/ui/input';
+import { validateField } from '~/lib/validation';
+
+interface LoginFormProps {
+  redirectAfterLogin?: boolean;
+  redirectPath?: string;
+}
+
+export default function LoginForm({
+  redirectAfterLogin = false,
+  redirectPath = '/',
+}: LoginFormProps) {
+  const { data: session, status } = useSession();
+  const [email, setEmail] = useState('');
+  const [password, setPassword] = useState('');
+  const [isLoading, setIsLoading] = useState(false);
+  const [error, setError] = useState('');
+  const [emailError, setEmailError] = useState('');
+
+  const router = useRouter();
+  const emailId = useId();
+  const passwordId = useId();
+
+  // 如果用户已登录，显示主应用内容
+  if (status === 'loading') {
+    return (
+      <div className="min-h-screen flex items-center justify-center">
+        <div className="text-center">
+          <div className="animate-spin rounded-full h-32 w-32 border-b-2 border-blue-600 mx-auto"></div>
+          <p className="mt-4 text-gray-600">加载中...</p>
+        </div>
+      </div>
+    );
+  }
+
+  if (status === 'authenticated' && session) {
+    return <Memo />;
+  }
+
+  const handleEmailChange = (value: string) => {
+    setEmail(value);
+    if (value) {
+      const validation = validateField('email', value);
+      setEmailError(validation.message);
+    } else {
+      setEmailError('');
+    }
+  };
+
+  const handleSubmit = async (e: React.FormEvent) => {
+    e.preventDefault();
+    setIsLoading(true);
+    setError('');
+
+    const emailValidation = validateField('email', email);
+    if (!emailValidation.isValid) {
+      setEmailError(emailValidation.message);
+      setIsLoading(false);
+      return;
+    }
+
+    try {
+      const result = await signIn('credentials', {
+        email,
+        password,
+        redirect: false,
+      });
+
+      if (result?.error) {
+        setError('邮箱或密码错误');
+      } else {
+        if (redirectAfterLogin) {
+          router.push(redirectPath);
+        } else {
+          // 登录成功后刷新页面状态，不需要跳转
+          router.refresh();
+        }
+      }
+    } catch (_error) {
+      setError('发生错误，请重试');
+    } finally {
+      setIsLoading(false);
+    }
+  };
+
+  return (
+    <div className="min-h-screen flex items-center justify-center bg-gradient-to-br from-slate-50 via-blue-50 to-indigo-100 py-12 px-4 sm:px-6 lg:px-8">
+      <div className="max-w-md w-full space-y-8">
+        <motion.div
+          initial={{ opacity: 0, y: -20 }}
+          animate={{ opacity: 1, y: 0 }}
+          className="text-center"
+        >
+          <motion.div
+            initial={{ scale: 0 }}
+            animate={{ scale: 1 }}
+            transition={{ delay: 0.2, type: 'spring', stiffness: 200 }}
+            className="w-20 h-20 bg-gradient-to-br from-blue-500 to-purple-600 rounded-full mx-auto mb-6 flex items-center justify-center"
+          >
+            <Sparkles className="w-10 h-10 text-white" />
+          </motion.div>
+          <h2 className="text-4xl font-bold bg-gradient-to-r from-blue-600 to-purple-600 bg-clip-text text-transparent">
+            登录账号
+          </h2>
+          <p className="mt-3 text-gray-600">
+            或者{' '}
+            <Link
+              href="/auth/signup"
+              className="font-medium text-blue-600 hover:text-blue-500 transition-colors"
+            >
+              注册新账号
+            </Link>
+          </p>
+        </motion.div>
+
+        <motion.form
+          initial={{ opacity: 0, y: 20 }}
+          animate={{ opacity: 1, y: 0 }}
+          transition={{ delay: 0.3 }}
+          className="mt-8 space-y-6"
+          onSubmit={handleSubmit}
+        >
+          <Card className="border-0 shadow-xl bg-white/80 backdrop-blur-sm">
+            <CardContent className="p-6 space-y-6">
+              <div className="space-y-2">
+                <label htmlFor={emailId} className="block text-sm font-medium text-gray-700">
+                  邮箱地址
+                </label>
+                <Input
+                  id={emailId}
+                  name="email"
+                  type="email"
+                  autoComplete="email"
+                  required
+                  error={!!emailError}
+                  leftIcon={<Mail className="w-4 h-4" />}
+                  placeholder="请输入邮箱地址"
+                  value={email}
+                  onChange={(e) => handleEmailChange(e.target.value)}
+                />
+                {emailError && (
+                  <motion.p
+                    initial={{ opacity: 0, y: -10 }}
+                    animate={{ opacity: 1, y: 0 }}
+                    className="text-sm text-red-600 flex items-center gap-1"
+                  >
+                    <span className="w-1 h-1 bg-red-600 rounded-full"></span>
+                    {emailError}
+                  </motion.p>
+                )}
+              </div>
+
+              <div className="space-y-2">
+                <label htmlFor={passwordId} className="block text-sm font-medium text-gray-700">
+                  密码
+                </label>
+                <Input
+                  id={passwordId}
+                  name="password"
+                  type="password"
+                  autoComplete="current-password"
+                  required
+                  leftIcon={<Lock className="w-4 h-4" />}
+                  placeholder="请输入密码"
+                  value={password}
+                  onChange={(e) => setPassword(e.target.value)}
+                />
+              </div>
+
+              {error && (
+                <motion.div
+                  initial={{ opacity: 0, scale: 0.95 }}
+                  animate={{ opacity: 1, scale: 1 }}
+                  className="bg-red-50 border border-red-200 rounded-lg p-3 text-red-600 text-sm text-center"
+                >
+                  {error}
+                </motion.div>
+              )}
+
+              <Button
+                type="submit"
+                disabled={isLoading || !!emailError}
+                loading={isLoading}
+                variant="gradient"
+                size="lg"
+                className="w-full group"
+              >
+                {isLoading ? (
+                  '登录中...'
+                ) : (
+                  <>
+                    登录
+                    <ArrowRight className="ml-2 w-4 h-4 group-hover:translate-x-1 transition-transform" />
+                  </>
+                )}
+              </Button>
+            </CardContent>
+          </Card>
+        </motion.form>
+      </div>
+    </div>
+  );
+}
